ai/requirements.txt
numpy
python-dotenv
google-genai
beautifulsoup4
yfinance
trafilatura
feedparser
requests
tenacity
pydantic
openai
deepeval
pytest
pytest-cov
vcrpy
pyyaml
asyncio
aiohttp
